requirements.txt
letta>=0.3.0
google-generativeai>=0.3.0
groq>=0.4.0
redis>=4.5.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
aioredis>=2.0.0
asyncio-mqtt>=0.13.0
